refactor(common): add plural on
